UIU-1916-add-dropdown-show-shortcuts centralize modal component
